refactor: 초대받지 않았을때 표시할 컴포넌트를 따로 분리
