+++ components/table/myInvitedDashboardTable/MyInvitedDashboardTable.module.scss
@@ -13,15 +13,6 @@
   line-height: normal;
 }
 
-.noInvitation {
-  @include column-flexbox();
-  @include text-style-18-light();
-  width: 100%;
-  gap: 2.4rem;
-  color: $gray-400;
-  margin-top: 6.6rem;
-}
-
 .tableArea {
   margin-top: 20px;
   @include column-flexbox(start, stretch);
