+++ components/table/myInvitedDashboardTable/NoInvitation.module.scss
@@ -1,0 +1,8 @@
+.noInvitation {
+  @include column-flexbox();
+  @include text-style-18-light();
+  width: 100%;
+  gap: 2.4rem;
+  color: $gray-400;
+  margin-top: 6.6rem;
+}
